op-eventer/requirements.txt
google.cloud.bigquery
